5260 Split out new UploadDockerImagesToEcr
