+++ java/clients/src/main/java/sleeper/clients/admin/properties/AdminClientPropertiesStore.java
@@ -21,8 +21,8 @@ import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 import software.amazon.awssdk.services.s3.S3Client;
 
 import sleeper.clients.deploy.container.DockerImageConfiguration;
-import sleeper.clients.deploy.container.UploadDockerImagesToEcr;
 import sleeper.clients.deploy.container.UploadDockerImagesToEcrRequest;
+import sleeper.clients.deploy.container.UploadDockerImagesToEcrTemp;
 import sleeper.clients.util.ClientUtils;
 import sleeper.clients.util.cdk.CdkCommand;
 import sleeper.clients.util.cdk.InvokeCdkForInstance;
@@ -54,12 +54,12 @@ public class AdminClientPropertiesStore {
     private final DynamoDbClient dynamoClient;
     private final InvokeCdkForInstance cdk;
     private final DockerImageConfiguration dockerImageConfiguration;
-    private final UploadDockerImagesToEcr uploadDockerImages;
+    private final UploadDockerImagesToEcrTemp uploadDockerImages;
     private final Path generatedDirectory;
 
     public AdminClientPropertiesStore(
             S3Client s3Client, DynamoDbClient dynamoClient, InvokeCdkForInstance cdk,
-            Path generatedDirectory, UploadDockerImagesToEcr uploadDockerImages) {
+            Path generatedDirectory, UploadDockerImagesToEcrTemp uploadDockerImages) {
         this.s3Client = s3Client;
         this.dynamoClient = dynamoClient;
         this.dockerImageConfiguration = DockerImageConfiguration.getDefault();
