+++ java/clients/src/main/java/sleeper/clients/deploy/DeployExistingInstance.java
@@ -23,8 +23,8 @@ import software.amazon.awssdk.services.ecr.EcrClient;
 import software.amazon.awssdk.services.s3.S3Client;
 
 import sleeper.clients.deploy.container.EcrRepositoryCreator;
-import sleeper.clients.deploy.container.UploadDockerImagesToEcr;
 import sleeper.clients.deploy.container.UploadDockerImagesToEcrRequest;
+import sleeper.clients.deploy.container.UploadDockerImagesToEcrTemp;
 import sleeper.clients.deploy.jar.SyncJars;
 import sleeper.clients.util.ClientUtils;
 import sleeper.clients.util.cdk.CdkCommand;
@@ -112,7 +112,7 @@ public class DeployExistingInstance {
                 .deleteOldJars(false)
                 .build().sync();
 
-        UploadDockerImagesToEcr.builder()
+        UploadDockerImagesToEcrTemp.builder()
                 .baseDockerDirectory(scriptsDirectory.resolve("docker")).jarsDirectory(jarsDirectory)
                 .ecrClient(EcrRepositoryCreator.withEcrClient(ecr))
                 .build().upload(runCommand, UploadDockerImagesToEcrRequest.forExistingInstance(properties));
