+++ java/clients/src/main/java/sleeper/clients/deploy/DeployNewInstance.java
@@ -26,8 +26,8 @@ import software.amazon.awssdk.services.sts.StsClient;
 
 import sleeper.clients.deploy.container.EcrRepositoryCreator;
 import sleeper.clients.deploy.container.StackDockerImage;
-import sleeper.clients.deploy.container.UploadDockerImagesToEcr;
 import sleeper.clients.deploy.container.UploadDockerImagesToEcrRequest;
+import sleeper.clients.deploy.container.UploadDockerImagesToEcrTemp;
 import sleeper.clients.deploy.jar.SyncJars;
 import sleeper.clients.deploy.properties.PopulateInstancePropertiesAws;
 import sleeper.clients.table.AddTable;
@@ -136,7 +136,7 @@ public class DeployNewInstance {
         SyncJars.builder().s3(s3Client)
                 .jarsDirectory(jarsDirectory).instanceProperties(instanceProperties)
                 .deleteOldJars(false).build().sync();
-        UploadDockerImagesToEcr.builder()
+        UploadDockerImagesToEcrTemp.builder()
                 .baseDockerDirectory(scriptsDirectory.resolve("docker")).jarsDirectory(jarsDirectory)
                 .ecrClient(EcrRepositoryCreator.withEcrClient(ecrClient))
                 .build().upload(runCommand,
